config backbone
