+++ paddleseg/models/backbones/__init__.py
@@ -24,5 +24,4 @@ from .stdcnet import *
 from .lite_hrnet import *
 from .shufflenetv2 import *
 from .ghostnet import *
-from .hrnetv2_psa import *
 from .top_transformer import *
